backend/requirements.txt
fastapi[standard]
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart 
sqlalchemy 
pymysql
alembic
pydantic-settings==2.8.1